requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.2
python-pptx>=0.6.23
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.0.0
google-genai>=1.44.0
pydantic>=2.12.0